TG-20: src/index.js divided into index.js & app.js
